Create tass_python_3_client.py
Example of a Python 3 client to help those who are struggling to make the original file work with Python 3 and the Crypto package. I used the original file here so that people can upgrade their code with a minimal number of changes. There are also examples on how to call certain methods in the API.